requirements.txt
Django==3.0.7
django-heroku==0.3.1
django-multiselectfield==0.1.12
Pillow==7.2.0
gunicorn==20.0.4
django-storages==1.9.1
boto3==1.14.33